# Improve logql parser allocations.