--- a/pkg/util/conv.go
+++ b/pkg/util/conv.go
@@ -3,56 +3,23 @@ package util
 import (
 	"math"
 	"sort"
-	"strings"
 	"time"
 	"unsafe"
 
 	"github.com/cortexproject/cortex/pkg/ingester/client"
 	"github.com/prometheus/common/model"
 	"github.com/prometheus/prometheus/pkg/labels"
-
-	"github.com/grafana/loki/pkg/logql"
+	"github.com/prometheus/prometheus/promql/parser"
 )
-
-type byLabel []client.LabelAdapter
-
-func (s byLabel) Len() int           { return len(s) }
-func (s byLabel) Less(i, j int) bool { return strings.Compare(s[i].Name, s[j].Name) < 0 }
-func (s byLabel) Swap(i, j int)      { s[i], s[j] = s[j], s[i] }
 
 // ToClientLabels parses the labels and converts them to the Cortex type.
 func ToClientLabels(labels string) ([]client.LabelAdapter, error) {
-	matchers, err := logql.ParseMatchers(labels)
+	ls, err := parser.ParseMetric(labels)
 	if err != nil {
 		return nil, err
 	}
-	result := make([]client.LabelAdapter, 0, len(matchers))
-	for _, m := range matchers {
-		result = append(result, client.LabelAdapter{
-			Name:  m.Name,
-			Value: m.Value,
-		})
-	}
-	sort.Sort(byLabel(result))
-	return result, nil
-}
-
-// ParseLabels parses labels from a string using logql parser.
-func ParseLabels(lbs string) (labels.Labels, error) {
-	matchers, err := logql.ParseMatchers(lbs)
-	if err != nil {
-		return nil, err
-	}
-
-	result := make(labels.Labels, 0, len(matchers))
-	for _, m := range matchers {
-		result = append(result, labels.Label{
-			Name:  m.Name,
-			Value: m.Value,
-		})
-	}
-	sort.Sort(result)
-	return result, nil
+	sort.Sort(ls)
+	return client.FromLabelsToLabelAdapters(ls), nil
 }
 
 // ModelLabelSetToMap convert a model.LabelSet to a map[string]string